+++ Tests/MoyaProviderIntegrationTests.swift
@@ -224,10 +224,10 @@ class MoyaProviderIntegrationTests: QuickSpec {
                     var plugin: NetworkLoggerPlugin!
                     beforeEach {
                         log = ""
-
-                        plugin = NetworkLoggerPlugin(verbose: true, output: { printing in
+                        
+                        plugin = NetworkLoggerPlugin(verbose: true, output: { (separator: String, terminator: String, items: Any...) in
                             //mapping the Any... from items to a string that can be compared
-                            let stringArray: [String] = printing.2.map { $0 as? String }.flatMap { $0 }
+                            let stringArray: [String] = items.map { $0 as? String }.flatMap { $0 }
                             let string: String = stringArray.reduce("") { $0 + $1 + " " }
                             log += string
                         })