+++ Sources/Moya/MoyaProvider+Internal.swift
@@ -235,7 +235,7 @@ private extension MoyaProvider {
         }
 
         // Perform the actual request
-        if let _ = progressCompletion {
+        if progressCompletion != nil {
             switch progressAlamoRequest {
             case let downloadRequest as DownloadRequest:
                 if let downloadRequest = downloadRequest.downloadProgress(closure: progressClosure) as? T {